64-bit GDT Loaded

+++ src/kernel/arch/x86/isr.h
@@ -2,9 +2,9 @@
 #include <stdint.h>
 
 typedef struct {
-    uint64_t rdi, rsi, rbp, shit, rbx, rdx, rcx, rax;        // getting those shit regs from stack
+    uint64_t rdi, rsi, rbp, shit, rbx, rdx, rcx, rax;           // getting those shit regs from stack
     uint64_t interrupt, error;                                  // no need for comment, It's obvious
-    uint32_t eip, cs, eflags, esp, ss;                          // It's pushed to the stack by stupid CPU
+    uint64_t eip, cs, eflags, esp, ss;                          // It's pushed to the stack by stupid CPU
 } __attribute__((packed)) Pushed_Regs;
 
 typedef void (*ISRHandler)(Pushed_Regs* regs);
